Add and run a stub implementation of parser to cuDF conversion
This commit implements a skeleton of the ToCudf() method for the
ParserOutputDevice<BaseAction> class, following examples of CopyToHost()
method from the same class, and of OutputPrinter::ToCsv() method.

It also adds its use to `benchmark/main.cu` file, to test that it
works correctly, and benchmarks its use.  The dummy implementation
of ToCudf() takes the same magnitude of time as initialization,
which for sample_900000.json file is:

  META-JSON-PARSER
  Time measured by GPU:
  + Initialization:                             2240 ns
  [...]
  + Converting to cuDF format:                  7424 ns