Implementing CudfBoolColumn converter (not tested yet)
The rmm::device_buffer class, which is used by cudf::column to store data,
is closed to extensions, because all fields are private, and there are
setters missing for _size and _capacity fields.  All constructors, operators
and methods in rmm::device_buffer assume that we want to _copy_ the data from
CPU, not that we have data already on GPU that we don't want to copy.  This
required some hacking.

To be able to modify private fields without setters, we have used the
horrible hack, taking advantage of undefined behavior in union type
handling.  The code assumes that rmm::device_buffer would not change the
layout. TODO: add at least some sanity checks.

TODO: fix the add_column() function / CudfConverter::call() method
signature, to take number of elements and the size of single element.

Note that the result is not checked if it is correct, or at least well
formed.