benchmark: Implement to_csv_libcudf() to test cuDF conversion
The newly created function is used to test if conversion to cuDF
format (to cudf::table), residing on GPU, works correctly.  Currently
it overwrites the CSV file that was created with the OutputPrinter.

Running the ./meta-json-parser-benchmark we find out there is some
bug in the code, probably some data is owned by two different objects,
with different lifetimes:

terminate called after throwing an instance of 'cudf::logic_error'
  what():  cuDF failure at: [...]/cudf/cpp/src/column/column_view.cpp:53: Null data pointer.
Aborted