Implemented CudfUnknownColumnType converter
This class is now used on ParserOutputDevice::CudfColumnsConverterList
list instead of previously used std::false_type for column types for
which we do not know how to convert them to cudf::column, that is those
without CudfConverter inner type (that is, `using CudfConverter = ...`).

This would allow us in the next commit to remove the specialization
of the add_column function template.  The code was changed and tested
to check that add_column<std::false_type> specialization was no longer
used.  It would make possible to pass additional type metaparemeters
to add_column in future, if needed (function templates cannot be partially
specialized, as opposed to class templates).