run_benchmarks: Update examine_benchmarks.ipynb notebook
NOTE that this Jupyter notebook is saved as is, without any processing,
with internal state included.  What should be done to make it more
friendly to version control would be to use nbdev, or jupytext,
or a similar solution.

NOTE that the file with the data displayed is assumed to have generic
default 'benchmark.csv' filename.