benchmark: Prepare for libcudf (with conditional compilation)
NOTE: new start for adding libcudf support to the meta-json-parser.

Make it possible to configure meta-json-parser-benchmark to use libcudf
(as an optional deependency), and if the build process is configured
to use this library, then add appropriate build flags -- currently
hardcoded to values that make sense for the 'choinka' host.

The code simply includes appropriate #include-s for further use of
the libcudf library.  This means that libcudf code is not exercised,
for now.  We test that '#ifdef HAVE_LIBCUDF' (perhaps to be renamed
to USE_LIBCUDF) works, and that '#include <cudf/...>' works, but
we do not test that linker is configured correctly for libcudf use.