benchmark: Group subset of options into meta-json-parser group
This grouping of options with `.add_option_group(...)` is here
to make it easier to disable a set of options as a whole, or
for a set of options to require some other option.

See: https://github.com/CLIUtils/CLI11#option-groups

If only option grouping in the help output is needed, there is an
easier way, using `app.option_defaults()->group("<name>");`.