benchmark: Implement two versions of to_cvs_libcudf()
The code was copied from 4b55b0b50 and from the 'test-libcudf.cpp'
file in the https://github.com/mis-wut/test-libcudf project.

There are two versions: one that accepts table with metadata (like
the names of fields / columns), and one without.  The CSV file
generated by parsing with meta-json-parser and the one generated
by parsing with libcudf are identical.

TODO: make it possible to turn on printing of CSV header taking names
from metadata; currently the important part is turned off because
of strange compile error:

  benchmark/main.cu(940): error: argument of type "const cudf::io::table_metadata *"
  is incompatible with parameter of type "cudf::io::table_metadata *"