debug_helpers: Add type_id_to_name() and describe_column()
Those functions (currently ot documented using Doxygen) are used
to describe structure of cudf::table.  This would help with debugging
any issues with converting the metaparser output to this type.

The contents of cudf::table can be dumped (for some types of columns)
with newly added column_data_dumper().

NOTE: compile tested, but not that it works correctly (it should).