benchmark: Add --cudf-structure option
This option currently works only when using --use-libcudf-parser,
that is when using libcudf library to parse JSON.  It makes the
meta-json-parser-benchmark program print the structure of generated
cudf::table, or rather cudf::io::table_with_metadata.

Prinnting the structure of cudf::column and structure of cudf::table
will be useful when debugging conversion of metaparser results to
cuDF-compatibile output.