benchmark: Skeleton of converting metaparser results to cudf::table
This consist of adding the output_to_cudf(), though currently it creates
an empty table instead of converting metaparser output, and using it
in main_templated().  There is also a CUDA event created and registered
for benchmarking the conversion process, but it is not used yet.  It may
have to wait for the refactoring of checkpoint handling, that is of
print_results().

Note that the declaration of to_csv_libcudf() had to be moved earlier
to be able to use it in main_templated(), but currently it is not
used there.

While at it, correct indentation in print_results().