parser_output_device: Implement ToCudf(), iterating over fields
To make it work, we need both the converter type, and the output request,
so store the original type with the converter type as `OT`; we can then
extract the request itself from it.  The request is needed to find the
size of an element (following what is done in CopyToHost() method.

The CudfColumnConverterList is now created out of filtered list, out
of only those actions with output.