Add CudfNumericConverter to JNumber
This allows to convert JNumber fields to cudf::column, which then
will be composed into cudf::table.

Note that to convert metaparser result to cuDF-compatibile output
we need to have the result as rmm::device_buffer or something that
can be converted to rmm::device_vector, or something that
rmm::device_buffer can be constructed from with one-argument
constructor.  Currently it is done as horrible hack using union
aliasing, but in the future metaparser might be configured to use
rmm::device_buffer, rmm::device_vector or rmm::device_uvector
in place of currently used thrust::device_vector for storing data
on GPU... if we don't want to copy data GPU to GPU.

TODO: Remove debugging (via printing various values).