benchmark: Fix compilation error in 'enum end_of_line { unknown, unix, win };'
After updating the system, with the new GCC (most probably), the
benchmark/main.cu stopped compiling, with the following error:

  ../meta-json-parser/benchmark/main.cu(61): error: expected an identifier

The problem is with using `unix` as an identifier.  It is not a keyword
defined by the standard, but something that is common to find #defined
on Unix system, see
  https://stackoverflow.com/questions/3770322/is-unix-restricted-keyword-in-c

You can check this with the following command:
  $ echo | gcc -E -dD - | grep unix
  #define __unix 1
  #define __unix__ 1
  #define unix 1

This is with GCC 11.2.0 on Linux.

Because this problem is on the preprocessor levels, C++ namespacing
(using `enum class` instead of `enum`) does not help.  What is strange
adding the following also did not work:

  #ifdef unix
  #undef unix
  #endif

At the end, the name was changed from `end_of_line::unix` to
`end_of_line::uniks`.

.
While at it, change plain old `enum` to `enum class`, following the
C++ Core Guidelines - Enum.3: Prefer class enums over “plain” enums
http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#enum3-prefer-class-enums-over-plain-enums

This necessitated the slight change in the signature of the
workgroup_sizes_desc() function, and the minor modification of
main_templated().

While at it, document enum constants.