Add CudfBoolConverter to JBool
This allows to convert JBool fields to cudf::column, which then
will be composed into cudf::table.

NOTE that in cuDF, the only bool type is cudf::type_id::BOOL8,
so if JBool is used with first metaparameter which is not an 8-bit
type like uint8_t, then it would be converted instead into a numerical
column; this is different from the first attempt, where such JBool
was simply skipped.