Add CudfColumnConverter to JStringDynamicCopy{,V2,V3}
This allows to convert JStringDynamicCopy, JStringDynamicCopyV2 and
JStringDynamicCopyV3 fields to cudf::column, which then will be
composed into cudf::table.

The CudfDynamicStringColumn<type> is used to do the conversion.  If
everything goes as expected, it is simple wrapping of existing data
on GPU into a container, and it does not involve either allocating
or copying data on CPU or on GPU.

.
NOTE: there was a problem with cudf::type_to_id<char>() returning
cudf::type_id::EMPTY, instead of cudf::type_id::INT8 as expected,
which necessitated adding printing debugging information.  This
will be removed in the next commit.