benchmark: Begin rewrite of way of checkpointing results
The benchmarking part, that is measuring how long does it take to
perform specific operations, depends a bit on the options selected,
for example if error checking is performed.  When new operation
is to be added to be benchmarked, like the conversion of meta-parser
output to cudf::table, it would get more complicated, especially
if the operation is conditional on the command like option or compile
option like HAVE_LIBCUDF.

Instead of making `print_results()` even more of the mess of `if`-s
and `#ifdef`-s, lets rewrite the way checkpointing is done.  Let's
add each checkpoint to the list, and then use this list to print
results.

This commit creates the checkpoint_results.{h,cpp} source code files,
and the `checkpoint_event()` function.  The internal details of how
checkpoints are stored is kept hidden, encapsulated in the new module.
We start with duplicating the current implementation of `print_results()`,
with the goal of replacing it in the future.

One complication that is planned for (but not yet implemented) is
that some operations, like JSON parsing, are split into sub-operations,
like JSON processing and post kernel hooks.

To check that we caught all checkpoints there is `print_checkpoint_events()`
function...  It turned out that some events are checkpointed outside
'benchmark/main.cu', which would necessitate some changes in the code.

TODO:
- use checkpoint_event() everywhere cudaEventRecord is used
- add support for sub-operations
- reproduce the print_results() exactly

Note that this "refactoring" would enable to add other metrics to
checkpoints, like CPU time, or GPU memory.