benchmark: print_checkpoint_events() now prints accumulated time
Make `print_checkpoint_events()` more useful by printing automatically
computed total accumulated time measured by GPU, starting from the
first event.

NOTE: This lead to the discovery that event denoting group of sub-events
was put in the wrong place, which would result in preceding operation
to print wrongly that it takes 0 ns.  This would be fixed in the
following commit.