benchmark: Move "Parsing total" event to parser_kernel.cuh
This is one way of handling events that are to be sum of the
following sub-events.  Alternative would be to mark that it is
special kind of event by passing std::nullptr (or 0) as first
parameter to `checkpoint_event()`, but it would lead to a bit
more complicated processing side.

With this change we no longer need `gpu_parsing_start_checkpoint`
variable.