benchmark: Implement "sum of the following" in print_checkpoint_results()
Now events which are just a header for a set of sub-event / sub-operations,
also known as "summ of the following" type of events, now correctly print
the summary time of all sub-events.

Note that this time might differ from what print_results() from
benchmark/main.cu shows because instead of adding the timings of subevents
it measures the time from first sub-event to the event after all
sub-events (found with `find_next_main_event()`).