parser_output_device: Add my_strlen_s and to_cudf_pair
Add the `my_strnlen_s` function which can be used to compute the length
of the C string with given maximum length.

Add the `to_cudf_pair<maxCharacters>` unary functor which returns a pair
of C-string + length, where each string is assumed to be at most
maxCharacters long.  This uses `my_strnlen_s`.

The `to_cudf_pair` "function" will be used in CudfStaticStringConverter.