benchmark: Create null mask for dynamic string as empty buffer
Previously cudf::create_null_mask(n_elements, cudf::mask_state::UNALLOCATED)
was used instead.  This was different from other overloads for the
cudf::make_strings_column() function, which passed as a default value
for mask an empty rmm::device_buffer - like it is done now.

It might be the case that the problem was with lack of std::move around
the parameter; is it only for objects created with constructors passed as
lvalue params (passed via &&) that are copy-elided?

This changed the time it took to convert the result to `cudf::table`
decrease to the values of the same magnitude but smaller than post
kernel hooks.

Before:
```
+ Parsing total (sum of the following):   41711391 ns
  - JSON processing:                      40607425 ns
  - Post kernel hooks:                     1103968 ns
+ Copying output:                       1486018676 ns
+ Converting to cuDF format:             816141113 ns
```

After
```
+ Parsing total (sum of the following):   42864929 ns
  - JSON processing:                      40663391 ns
  - Post kernel hooks:                     2201535 ns
+ Copying output:                       1495105957 ns
+ Converting to cuDF format:               1962592 ns
```