run_benchmarks.py: Make it run also with click 7.x
This script was created to run with click 8.x; for some reason,
possibly because it has click 7.x installed, this script failed
to run inside the RAPIDS docker conntainer.

The problem was caused by the fact that the default value somehow
got to be bytes, not str, and sutomatic conversion to pathlib.Path
failed.  This commit adds and uses the ensure_path() function to
help solve this issue.

While at it, improve and expand script messages.