run_benchmarks.py: Ignore stderr of ./meta-json-parser-benchmark
Because of the horrible hack used to be able to convert Thrust vectors
to RMM vectors (something that cuDF can understand and use), there
is most likely confusion about the ownership of data, resulting in
double free after the end of the program.

Let's ignore error messages, like the following one, or rather redirect
them to /dev/null:

  terminate called after throwing an instance of 'thrust::system::system_error'
    what():  CUDA free failed: cudaErrorInvalidValue: invalid argument
  Aborted (core dumped)

that are printed to standard error (stderr).