run_benchmarks.py: Fix handling of --samples=<n> option
Before this commit each result for the same parameters was written to
the same object, instead of using a separate object (dict) for each
sample.  The result of this bug was that this script wrote <samples>
time the last result, instead of writing <samples> results for each
set of parameters.

This fixes this bug, though it looks like sometimes it writes empty
results for a few last entries.