benchmark: Print information about input file and expected objects
This information was planned to be used to be able to save the output
of ./meta-json-parser-benchmark command to a file, and then parse it
with run_benchmarks.py -- extracting the name of the JSON file and
the expected number of objects.

TODO: Parse this new info in run_benchmarks.py, extracting it.