benchmark: Add examples/ folder with JSON files and data_def.cuh's
In the `examples/` folder there are test files for checking that
./meta-json-parser-benchmark works correctly, each test case composed
of the JSON Lines *.json file and corresponding *.data_def.cuh file.
The latter needs to be copied as `data_def.cuh` to `benchmarks/`,
that is up-directory.  The original `data_def.cuh` for generated JSON
is stored as `sample.data_def.cuh`.

TODO: modify `poc/json2cuda.py` script in https://github.com/mis-wut/json2meta
so that appropriate `#include <meta_json_parser/action/*.cuh>` are
in `data_def.cuh` and not in `main.cu`.  With the proposed change
we can include only those headers that are needed.

This TODO assumes that it would not be possible to pre-compile headers.