Separate datetime options from JDatetime action to prevent circular dependencies
The access to JDatetimeOptions will be needed in the upcoming change
which will add support for converting datetime column to cuDF format;
the to be added CudfDatetimeColumn needs to know the resolution of
datetime data (is it counting seconds, or milliseconds, etc.).

Rebased to be a preparatory change (and avoid revisions which do not
compile).