benchmark: Add run_benchmarks_types.sh script, for testing parsing
This script was originally intended to make it easy to benchmark and
compare meta-parser vs libcudf when parsing a specific type of data:
floats, integers, strings, datetimes, etc.

Currently it can be used to check whether parsing with meta-parser
matches what libcudf parses.

NOTE: this script does not use generate_and_compile.sh (which
incorporates generating meta-definition based on a sample of the JSON
file to parse), and neither does it borrow any code from it.

Example of use:
---------------
$ run_benchmarks_types.sh examples/datetime.data_def.cuh \
                          examples/datetime.json