examine_benchmarks.ipynb: Analyze skipping fields
Check how meta-json-parser-benchmark behaves for different assumptions
about the order and structure of fields in JSON-Lines dicts.

The analysis includes comparison of running the parser with an
assumption that the order of fields in JSON dictionary is fixed, and
that we know all the fields, with subsequent relaxing of those
assumptions.

One of checks was how the performance behaves if we need to parse only
selected part of JSON fields.