benchmark/examples: Add nested-*depth=*.data_def.cuh
These are to be used for the nested-depth={n}_{size}.jsonl data
generated with the help of the generate_nested_pos.py script from
the json2meta repository, to examine how performance depends on
the nesting depth.

The nested-depth=5.jsonl file has the following structure:

  {"l1":{"l2":{"l3":{"l4":{"l5":800}}}}}

The nested-depth=1.jsonl file has the following structure:

  {"l1":800}

and is therefore shorter than the nested-depth=5.jsonl file.

The nested-fakedepth=5.jsonl file was created with the help of
$ cat nested-depth\=5_1000000.jsonl | \
  jq -c '{"l1_____l2_____l3_____l4_____l5":.l1.l2.l3.l4.l5}' \
  >nested-fakedepth\=5_1000000.jsonl

and has the following structure:

  {"l1_____l2_____l3_____l4_____l5":800}

which makes it the same length as the nested-depth=5.jsonl file.