CMakeLists.txt: Use $ENV{CONDA_PREFIX} for includes and libraries
But keep old hardcoded /opt/conda/envs/rapids/{include,lib} as is,
because it is used in docker image with RAPIDS (with libcudf), and
CONDA_PREFIX environment variable is not set for our docker image
(is not defined in our Dockerfile).

This is 3rd part of split commit 1c9152f "Make all libraries local
and fix include errors".